Small improvement to stabs reader to avoid warnings of this form:
--14081-- warning: function CloseSocket__Fi:F(0,20) missing closing N_FUN stab at entry 10052

which were common when compiling with GCC's -gstabs option.

Instead of relying on an end-of-function N_FUN entry to handle the final
N_SLINE of a function, it can now handle it by finding the difference between
the start of the previous function and the start of the next.

Not extensively tested, but I think it works ok.


git-svn-id: svn://svn.valgrind.org/valgrind/trunk@386 a5019735-40e9-0310-863c-91ae7b9d1cf9