Partially merge in a (heavily modified) patch from Eric Estievenart which adds support for reading directory names from DWARF2 debug info.
Also rework the representation of file and directory tables in the
DWARF2 reader.  This removes a longstanding but only-just-discovered
curiousity that the previous code expanded the filename table one
entry at a time, so that reading file names from a DWARF2 object was
quadratic in the number of file names.  It's now N log N.




git-svn-id: svn://svn.valgrind.org/valgrind/trunk@3908 a5019735-40e9-0310-863c-91ae7b9d1cf9