Handle the memory written by aio_read() properly -- mark the memory buffer as written once aio_return() is successfully called.
Also check the addressability of the buffer for both aio_read() and
aio_write().

Also check the file descriptor for aio_read() and aio_write().

And add a test for this.  There's one corner case of the test that doesn't
work as expected and is currently commented out.  But aio_*() certainly
works better than it used to.

All this is for bug 197227.


git-svn-id: svn://svn.valgrind.org/valgrind/trunk@10539 a5019735-40e9-0310-863c-91ae7b9d1cf9